requirements.txt
beautifulsoup4==4.12.2
EbookLib==0.18
langchain==0.0.175
libgen_api==1.0.0
openai==0.27.2
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.28.1
